correct res check from previous commit

+++ checker/check_resolution.c
@@ -3,8 +3,8 @@
 
 void	verify_max_screen_size(t_window_settings window, t_cub *config)
 {
-	long int	current_x_res;
-	long int	current_y_res;
+	int	current_x_res;
+	int	current_y_res;
 
 	mlx_get_screen_size(window.mlx, &current_x_res, &current_y_res);
 	if (config->x_res > current_x_res)
